requirements.txt
openai
python-dotenv
rich
fastapi
uvicorn
